re-organized structs.
