+++ src/adapter/migration_handler.go
@@ -1,0 +1,9 @@
+package adapter
+
+type MigrationHandler interface {
+	GetVersion() uint64
+	GetType() string
+	GetName() string
+	Up() error
+	Down() error
+}
